moved update delegation
